requirements.txt
requests
pandas
scikit-learn
googlesearch-python
colorama
beautifulsoup4